[vsts-938452] Change inheritance order of altool-task classes
