+++ msbuild/Xamarin.Mac.Tasks.Core/Xamarin.Mac.Tasks.Core.csproj
@@ -38,6 +38,8 @@
     <Compile Include="Tasks\MmpTaskBase.cs" />
     <Compile Include="Tasks\XamMacArch.cs" />
     <Compile Include="Tasks\ArchiveTaskBase.cs" />
+    <Compile Include="Tasks\ALToolUploadTaskBase.cs" />
+    <Compile Include="Tasks\ALToolValidateTaskBase.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\external\Xamarin.MacDev\Xamarin.MacDev\Xamarin.MacDev.csproj">
