+++ msbuild/Xamarin.Mac.Tasks/Tasks/ALToolUpload.cs
@@ -1,21 +1,6 @@
-using System.Text;
-
-using Xamarin.MacDev;
-using Xamarin.MacDev.Tasks;
-
 namespace Xamarin.Mac.Tasks
 {
-	public class ALToolUpload : ALToolTaskBase
+	public class ALToolUpload : ALToolUploadTaskBase
 	{
-		public override PlatformName FileType => PlatformName.MacOSX;
-		
-		protected override string GenerateCommandLineCommands ()
-		{
-			var sb = new StringBuilder ();
-			sb.Append ("--upload-app ");
-			sb.Append (base.GenerateCommandLineCommands ());
-
-			return sb.ToString ();
-		}
 	}
 }
