+++ msbuild/Xamarin.Mac.Tasks/Tasks/ALToolVerify.cs
@@ -1,21 +1,6 @@
-using System.Text;
-
-using Xamarin.MacDev;
-using Xamarin.MacDev.Tasks;
-
 namespace Xamarin.Mac.Tasks
 {
-	public class ALToolValidate : ALToolTaskBase
+	public class ALToolValidate : ALToolValidateTaskBase
 	{
-		public override PlatformName FileType => PlatformName.MacOSX;
-		
-		protected override string GenerateCommandLineCommands ()
-		{
-			var sb = new StringBuilder ();
-			sb.Append ("--validate-app ");
-			sb.Append (base.GenerateCommandLineCommands ());
-
-			return sb.ToString ();
-		}
 	}
 }
